refactor(inference): one canonical model-format detector
Move ModelFormat/detect_format (plus the two format-specific error
message helpers) from lattice.rs's private backend module into a new
crates/inference/src/model_format.rs library module, documented
unstable/internal-binaries-only. lattice.rs, lattice_serve.rs, and
chat_metal.rs now all match on the shared Safetensors|Q4|Unknown enum
instead of each carrying its own boolean heuristic (detect_q4 in
lattice_serve.rs, is_q4_dir plus a partial safetensors-absence
re-check in chat_metal.rs) -- both deleted. No caller re-checks
sentinel files after matching on the enum.

The moved detector unit tests live in the new module; one test
(unreadable-dir fail-closed) was added for stronger coverage.

Amends docs/adr/ADR-080-consolidation-duplicated-contracts.md with an
implementation-record entry and a short amendment section covering the
new library-exposed helper (Status stays Accepted).

Closes #829
